requirements.txt
amqp==2.6.1
asgiref==3.4.1
billiard==3.6.4.0
cached-property==1.5.2
celery==4.3.0
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
django-celery-results==1.1.2
django-rest-framework==0.1.0
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.8.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==4.6.11
MarkupSafe==2.0.1
packaging==21.3
psycopg2==2.9.2
pyparsing==3.0.6
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
vine==1.3.0
zipp==3.6.0
